+++ src/components/pages/root/index.tsx
@@ -9,76 +9,80 @@ export const Index = async () => {
   const session = await getServerAuthSession();
 
   return (
-    <>
-      <h1 className="text-5xl font-extrabold tracking-tight sm:text-[5rem]">
-        Create <span className="text-[hsl(280,100%,70%)]">T3</span> App
-      </h1>
-      <div className="grid grid-cols-1 gap-4 sm:grid-cols-2 md:gap-8">
-        <Link
-          className="flex max-w-xs flex-col gap-4 rounded-xl bg-white/10 p-4 hover:bg-white/20"
-          href="https://create.t3.gg/en/usage/first-steps"
-          target="_blank"
-        >
-          <h3 className="text-2xl font-bold">First Steps →</h3>
-          <div className="text-lg">
-            Just the basics - Everything you need to know to set up your
-            database and authentication.
+    <main className="flex min-h-screen flex-col items-center justify-center bg-gradient-to-b from-[#2e026d] to-[#15162c] text-white">
+      <div className="container flex flex-col items-center justify-center gap-12 px-4 py-16">
+        <>
+          <h1 className="text-5xl font-extrabold tracking-tight sm:text-[5rem]">
+            Create <span className="text-[hsl(280,100%,70%)]">T3</span> App
+          </h1>
+          <div className="grid grid-cols-1 gap-4 sm:grid-cols-2 md:gap-8">
+            <Link
+              className="flex max-w-xs flex-col gap-4 rounded-xl bg-white/10 p-4 hover:bg-white/20"
+              href="https://create.t3.gg/en/usage/first-steps"
+              target="_blank"
+            >
+              <h3 className="text-2xl font-bold">First Steps →</h3>
+              <div className="text-lg">
+                Just the basics - Everything you need to know to set up your
+                database and authentication.
+              </div>
+            </Link>
+            <Link
+              className="flex max-w-xs flex-col gap-4 rounded-xl bg-white/10 p-4 hover:bg-white/20"
+              href="https://create.t3.gg/en/introduction"
+              target="_blank"
+            >
+              <h3 className="text-2xl font-bold">Documentation →</h3>
+              <div className="text-lg">
+                Learn more about Create T3 App, the libraries it uses, and how
+                to deploy it.
+              </div>
+            </Link>
           </div>
-        </Link>
-        <Link
-          className="flex max-w-xs flex-col gap-4 rounded-xl bg-white/10 p-4 hover:bg-white/20"
-          href="https://create.t3.gg/en/introduction"
-          target="_blank"
-        >
-          <h3 className="text-2xl font-bold">Documentation →</h3>
-          <div className="text-lg">
-            Learn more about Create T3 App, the libraries it uses, and how to
-            deploy it.
-          </div>
-        </Link>
-      </div>
-      <div className="flex flex-col items-center gap-2">
-        <p className="text-2xl text-white">
-          {hello ? hello.greeting : "Loading tRPC query..."}
-        </p>
-
-        <div className="flex flex-col items-center justify-center gap-4">
-          {session !== null && (
-            <p className="text-center text-2xl text-white">
-              <span>Logged in as {session.user?.name}</span>
+          <div className="flex flex-col items-center gap-2">
+            <p className="text-2xl text-white">
+              {hello ? hello.greeting : "Loading tRPC query..."}
             </p>
-          )}
 
-          {session === null && (
-            <Link
-              href="/api/auth/signin"
-              className="rounded-full bg-white/10 px-10 py-3 font-semibold no-underline transition hover:bg-white/20"
-            >
-              Sign in
-            </Link>
-          )}
+            <div className="flex flex-col items-center justify-center gap-4">
+              {session !== null && (
+                <p className="text-center text-2xl text-white">
+                  <span>Logged in as {session.user?.name}</span>
+                </p>
+              )}
 
-          {session !== null && !session.user.isAnonymous && (
-            <Link
-              href="/api/auth/signout"
-              className="rounded-full bg-white/10 px-10 py-3 font-semibold no-underline transition hover:bg-white/20"
-            >
-              Sign out
-            </Link>
-          )}
+              {session === null && (
+                <Link
+                  href="/api/auth/signin"
+                  className="rounded-full bg-white/10 px-10 py-3 font-semibold no-underline transition hover:bg-white/20"
+                >
+                  Sign in
+                </Link>
+              )}
 
-          {session !== null && (
-            <Link
-              href="/user"
-              className="rounded-full bg-white/10 px-10 py-3 font-semibold no-underline transition hover:bg-white/20"
-            >
-              My Page
-            </Link>
-          )}
-        </div>
+              {session !== null && !session.user.isAnonymous && (
+                <Link
+                  href="/api/auth/signout"
+                  className="rounded-full bg-white/10 px-10 py-3 font-semibold no-underline transition hover:bg-white/20"
+                >
+                  Sign out
+                </Link>
+              )}
+
+              {session !== null && (
+                <Link
+                  href="/user"
+                  className="rounded-full bg-white/10 px-10 py-3 font-semibold no-underline transition hover:bg-white/20"
+                >
+                  My Page
+                </Link>
+              )}
+            </div>
+          </div>
+
+          {session !== null && <LatestPost />}
+        </>
       </div>
-
-      {session !== null && <LatestPost />}
-    </>
+    </main>
   );
 };
